# Optimize building message packet