This PR is based on #1.

```
benchmark            old ns/op     new ns/op     delta
BenchmarkWrite-4     7470          2450          -67.20%

benchmark            old allocs     new allocs     delta
BenchmarkWrite-4     102            19             -81.37%

benchmark            old bytes     new bytes     delta
BenchmarkWrite-4     3600          1360          -62.22%
```