@methane 2016-07-19
There are some major points remaining.
Could you try them?

1. Use bytes.IndexByte instead of bytes.Split.
2. Unify bytes.Buffer.  You can build outer header in-place.
3. Cache buffer in RLogger.  Guard it with mutex.